add test for list command, add utils for generating strings and hex colors

+++ tests/test_commands.py
@@ -2,14 +2,23 @@ import pytest
 import subprocess
 import argparse
 import os
+import random
+import re
 from sjournal import SJournal, Note
-from utils_test import backup_file, delete_file, get_project_root, validate_note, validate_config
+from utils_test import backup_file, delete_file, \
+    get_project_root, \
+    validate_note, validate_config, \
+    random_string, random_hex_color
 from logger import logger
 
 # Suite of tests to validate the CLI interface for sjournal using subprocess to call the application
 
 ROOT_DIR = get_project_root()
 sjournal_py = f"{os.path.join(ROOT_DIR, 'sjournal.py')}"
+
+n_random_notes = 21
+n_random_categories = 3
+n_random_styles = 5
 
 
 @pytest.fixture(scope="function")
@@ -22,10 +31,11 @@ def clean_journal_fixture():
 
     # if "automated_test.db" exists, delete it
     journal_file = os.path.join(ROOT_DIR, "journals", "automated_test.db")
-    delete_file(journal_file)
+    if os.path.isfile(journal_file):
+        delete_file(journal_file)
 
     # create new test notebook
-    args = argparse.Namespace(command="load", journal_name="automated_test")
+    args = argparse.Namespace(command="load", journal_name="automated_test", debug=False)
     journal = SJournal(args)
     journal.run()
 
@@ -46,7 +56,34 @@ def journal_with_notes_fixture(clean_journal_fixture):
 
     for i in range(3):
         journal.create_connection()
-        journal.args = argparse.Namespace(command="add", category="General", content=[f"Note {i}"], style="")
+        journal.args = argparse.Namespace(command="add", category="General", content=[f"Note {i}"], style="", debug=False)
+        journal.run()
+
+    yield journal
+
+
+@pytest.fixture(scope="function")
+def random_journal(clean_journal_fixture):
+    journal = clean_journal_fixture
+
+    # Generate random category names
+    categories = []
+    for c in range(n_random_categories):
+        categories.append(random_string(10))
+    logger.debug(f"generated categores: {categories}")
+
+    # Generate random styles
+    styles = []
+    for s in range(n_random_styles):
+        styles.append(f"bold {random_hex_color()}")
+    logger.debug(f"generated styles: {styles}")
+
+    # Populate journal
+    for n in range(n_random_notes):
+        category = categories[n % len(categories)]
+        journal.create_connection()
+        content = ' '.join([random_string(random.randint(1, 20)) for i in range(50)])
+        journal.args = argparse.Namespace(command="add", category=category, content=[f"{content}"], style=random.choice(styles), debug=False)
         journal.run()
 
     yield journal
@@ -177,9 +214,83 @@ def test_delete_note(journal_with_notes_fixture):
             validate_note(journal.notes[-1], expected)
 
 
-def test_list():
-    pytest.skip("TODO")
+# quantity              Specify the amount of results to list
+# -a, --all             List all notes under given criteria
+# -c [CATEGORY], --category [CATEGORY] Choose a category of notes to list
+# -r, --reverse         Display notes in reverse chronological order
+# Start with a journal that contains a few notes
+@pytest.mark.parametrize('command', [
+        '',
+        'list',
+        'list 10',
+        'list -a',
+        'list -r',
+        'list -c',
+        'list 10 -a',
+        'list 10 -r',
+        'list 10 -c',
+        'list -a -r',
+        'list -a -c',
+        'list -r -c',
+        'list 10 -a -r',
+        'list 10 -a -c',
+        'list 10 -r -c',
+        'list -a -r -c',
+        'list 10 -a -r -c',
+    ]
+ )
+def test_list_default(random_journal, command):
 
+    journal = random_journal
+    notes = journal.notes
+
+    commandline = f"python {sjournal_py} --debug " + command
+
+    n_printed = 5
+    if '-c' in command:
+        # Get one category and determine the number of notes and number to print. will print 5 if more than 5 notes
+        all_categories = list(set([note.category for note in notes]))
+        commandline += f' "{all_categories[0]}"'
+        notes = [note for note in notes if note.category == all_categories[0]]
+        if len(notes) < 5:
+            n_printed = len(notes)
+    if '10' in command:
+        n_printed = 10
+    if '-a' in command:
+        n_printed = len(notes)
+
+    # List notes via commandline with debug option
+    logger.info(f"List notes via commandline")
+    logger.debug(f"commandline: {commandline}")
+    logger.debug(f"For this command, expect {n_printed} notes")
+
+    result = subprocess.run(commandline, capture_output=False)
+    logger.debug(result)
+    assert result.returncode == 0
+
+    # Read debug text into string
+    debug_file = os.path.join(ROOT_DIR, "reports", "debug.log")
+    with open(debug_file, "r") as output_file:
+        full_text = output_file.read()
+
+    logger.debug(full_text)
+
+    # Search output text for table headers
+    header = rf"ID\s+\|\s+Timestamp\s+\|\s+Category\s+\|\s+Content\s+"
+    match = re.search(header, full_text)
+    assert match, f"Could not find header in debug output"
+
+    # Search output text for each note
+    for i, note in enumerate(notes[-n_printed:]):
+        logger.info(f"SEARCHING FOR NOTE {i} (not ID) OUT OF {n_printed}")
+        regex = rf"{note.id}.*{note.timestamp}.*{note.category}.*{note.content[14:24]}"
+        match = re.search(regex, full_text)
+        assert match, f"Could not find Note # {note.id} in debug output\n{regex}"
+        logger.debug(match.group(0))
+
+
+def test_list_by_category(random_journal):
+    pytest.skip("TODO")
 
 def test_categories():
     pytest.skip("TODO")
